+++ Src/PerformanceCollector/Perf.Shared/Implementation/PerformanceCounterUtility.cs
@@ -23,9 +23,12 @@
         private const string Win32ProcessCategoryName = "Process";
         private const string ClrProcessCategoryName = ".NET CLR Memory";
         private const string Win32ProcessCounterName = "ID Process";
-        private const string ClrProcessCounterName = "Process ID";
-
+        private const string ClrProcessCounterName = "Process ID";        
+#if NETSTANDARD2_0
+        private const string StandardSdkVersionPrefix = "pccore:";
+#else
         private const string StandardSdkVersionPrefix = "pc:";
+#endif
         private const string AzureWebAppSdkVersionPrefix = "azwapc:";
         private const string AzureWebAppCoreSdkVersionPrefix = "azwapccore:";
 
@@ -136,7 +139,7 @@
         {
             if (IsWebAppRunningInAzure())
             {
-#if NETSTANDARD1_6
+#if NETSTANDARD1_6 || NETSTANDARD2_0
                 return AzureWebAppCoreSdkVersionPrefix;
 #else
                 return AzureWebAppSdkVersionPrefix;
@@ -264,7 +267,7 @@
 
         internal static string GetInstanceForCurrentW3SvcWorker()
         {
-#if NETSTANDARD1_6
+#if NETSTANDARD1_6 || NETSTANDARD2_0
             string name = new AssemblyName(Assembly.GetEntryAssembly().FullName).Name;
 #else
             string name = AppDomain.CurrentDomain.FriendlyName;
@@ -417,9 +420,13 @@
             {
                 // something went wrong and the category hasn't been found
                 // we can't perform this operation
+#if NETSTANDARD2_0
+                return Array.Empty<string>();
+#else
                 return new string[] { };
+#endif
             }
         }
 #endif
-    }
+            }
 }