+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Changelog
 
+## Version 2.10.0-beta3
+- Added NetStandard2.0 Target for PerfCounter project.
+- Added support for PerfCounters for .Net Core Apps in Windows.
+- Updated Base SDK to 2.10.0-beta3
+
 ## Version 2.10.0-beta2
-Updated Base SDK to 2.10.0-beta2
+- Updated Base SDK to 2.10.0-beta2
 
 ## Version 2.9.1
 - Updates Base SDK to version 2.9.1